poulailler-console/requirements.txt
Flask==1.1.1
RPi.GPIO=0.7.0
python-dotenv=0.10.5